efb_wechatcom_slave/requirements.txt
ehforwarderbot==2.1.1
lxml==4.9.1
python_magic==0.4.27
PyYAML==6.0
requests==2.28.1
websockets==10.3
